backend/requirements.txt
asn1crypto==0.24.0
click==7.1.2
crcmod==1.7
cryptography==2.6.1
distro-info==0.21
entrypoints==0.3
Flask==1.1.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.1
mysqlclient==2.0.1
pycrypto==2.6.1
PyGObject==3.30.4
python-apt==1.8.4.1
python-dotenv==0.14.0
pyxdg==0.25
SecretStorage==2.3.1
six==1.12.0
unattended-upgrades==0.1
Werkzeug==1.0.1
